src/main/java/com/yehor/phonecontacts/api/controller/AuthenticationController.java
package com.yehor.phonecontacts.api.controller;

import com.yehor.phonecontacts.api.dto.AuthenticationRequest;
import com.yehor.phonecontacts.api.dto.AuthenticationResponse;
import com.yehor.phonecontacts.api.dto.RegisterRequest;
import com.yehor.phonecontacts.service.AuthenticationService;
import lombok.RequiredArgsConstructor;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.PostMapping;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RestController;

@RestController
@RequiredArgsConstructor
public class AuthenticationController {

    private final AuthenticationService authenticationService;

    @PostMapping("/register")
    public ResponseEntity<AuthenticationResponse> register(@RequestBody RegisterRequest request) {
        return ResponseEntity.ok(authenticationService.register(request));
    }

    @PostMapping("/auth")
    public ResponseEntity<AuthenticationResponse> authenticate(@RequestBody AuthenticationRequest request) {
        return ResponseEntity.ok(authenticationService.authenticate(request));
    }
}
